third_party/WebKit/LayoutTests/inspector/coverage/coverage-view-filter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="coverage-test.js"></script>
<script src="resources/coverage.js"></script>

<link rel="stylesheet" type="text/css" href="resources/highlight-in-source.css">

<script>
async function test()
{
    InspectorTest.startCoverage();
    await InspectorTest.evaluateInPagePromise("performActions()");
    await InspectorTest.stopCoverage();

    var coverageView = self.runtime.sharedInstance(Coverage.CoverageView);
    setFilter("inspector");
    InspectorTest.dumpCoverageListView();
    setFilter("CES/COV");
    InspectorTest.dumpCoverageListView();
    setFilter("no pasaran");
    InspectorTest.dumpCoverageListView();
    setFilter("");
    InspectorTest.dumpCoverageListView();
    InspectorTest.completeTest();

    function setFilter(text)
    {
        coverageView._filterInput.setValue(text);
        coverageView._filterInput._onChangeCallback();
        InspectorTest.addResult(`Filter: ${text}`);
    }
}

</script>
</head>

<body onload="runTest()">
<p class="class">
Tests the filter is properly applied to coverage list view.
</p>

</body>
</html>
